Move mock files out of mock directory
