--- testing/mock/mod.ts
@@ -1,3 +1,0 @@
-/** This module is browser compatible. */
-
-export * from "./mock.ts";
